Separate parent functionality from ParentPollerWindows
ParentPollerWindows contained a mixture of methods to run in the parent
and the child, which was revealed by test failures now that
jupyter_client and ipython_kernel are separate packages. This moves the
parent methods into a new module jupyter_client.win_interrupt

Closes gh-8201
