+++ jupyter_client/manager.py
@@ -381,8 +381,8 @@ class KernelManager(ConnectionFileMixin):
         """
         if self.has_kernel:
             if sys.platform == 'win32':
-                from .parentpoller import ParentPollerWindows as Poller
-                Poller.send_interrupt(self.kernel.win32_interrupt_event)
+                from .win_interrupt import send_interrupt
+                send_interrupt(self.kernel.win32_interrupt_event)
             else:
                 self.kernel.send_signal(signal.SIGINT)
         else:
